replaced from reducer utils with test and action-creator with test
