+++ src/const/action.js
@@ -1,0 +1,5 @@
+export default {
+  CHANGE_CITY: `CHANGE_CITY`,
+  LOAD_OFFERS: `LOAD_OFFERS`,
+  AUTHORIZE: `AUTHORIZE`
+};
