+++ src/utils.js
@@ -1,0 +1,1 @@
+export const getOffersByCity = (selectedCity, offers) => offers.filter((offer) => offer.city.name === selectedCity.name);
